graph-rag-movielens/requirements.txt
fastapi>=0.111.0,<0.112.0
uvicorn[standard]==0.23.2
pydantic>=2.7.4,<3.0.0
langchain==0.3.2
langchain-ollama>=0.2.0,<0.3.0
qdrant-client==1.8.2
neo4j==5.11.0
numpy==1.26.2
pandas==2.1.2
tqdm==4.66.1
python-dotenv==1.0.0
requests==2.32.0